services/files-manager-service/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
psycopg[binary]==3.2.4
boto3==1.35.90
python-multipart==0.0.20
PyJWT==2.10.1
